Move all Proof API imports into proof and tasks packages
