--- src/main/java/com/sphereon/alfresco/blockchain/agent/backend/commands/controllers/FutureSuccess.java
@@ -1,9 +1,0 @@
-package com.sphereon.alfresco.blockchain.agent.backend.commands.controllers;
-
-import com.google.common.annotations.GwtCompatible;
-import org.checkerframework.checker.nullness.qual.Nullable;
-
-@GwtCompatible
-public interface FutureSuccess<V> {
-    void onSuccess(@Nullable V value);
-}
