--- src/main/java/com/sphereon/alfresco/blockchain/agent/backend/tasks/Task.java
@@ -1,5 +1,0 @@
-package com.sphereon.alfresco.blockchain.agent.backend.tasks;
-
-public interface Task<R> {
-    R execute();
-}
