+++ src/main/java/com/sphereon/alfresco/blockchain/agent/tasks/Task.java
@@ -1,0 +1,5 @@
+package com.sphereon.alfresco.blockchain.agent.tasks;
+
+public interface Task<R> {
+    R execute();
+}
